Improve Microsoft Graph error handling
If the wrong credentials or permissions are set for a Microsoft Graph
inbox, MailRoom will go into a tight loop of retrying the request to no
avail. https://docs.microsoft.com/en-us/graph/errors lists all the
possible errors. In most cases, we just want to terminate with a hard
failure since there is something the admin needs to check.